Add files via upload
We were given 2 lists, Students and sandwiches both containing 1 and 0, 1 for square-shaped sandwiches and 0 for circular-shaped sandwiches. Each student value denotes the type of sandwich they want in students list and the boolean value in sandwiches denotes the type of sandwich at the top of the stack; here the top of the stack is sandwiches[0] and students is a queue where the top or beginning of the queue is students[0]. We were to check how many students would go hungry and not eat the sandwiches because the sandwich at the top of the stack didnt match their preference.