Add files via upload
There were a set of operations specified that are as follows :
1.   ../ - Move to the parent directory of the current folder. If already in the parent directory do nothing.
2.   ./ - Remain in the same folder and do nothing.
3.  x/ - Move to the directory named x
We were given a list logs containing the log actions of a user. It was assumed that the user was already in the main folder and any action they took was inside of the main folder. With that in mind we were to find out the minimum number of operations it would take to get back to the main folder. For example lets consider a log : logs = logs = ["d1/","d2/","../","d21/","./"], in this case the log cycle would look something like this :
Goto d1 -> goto d2 -> goto parent folder of d2 == d1-> go to d21 -> remain in d21; or in other words d1 - d2 - d1 -d21 that is it would take us 2 operations to get back to the main folder. 1 operation to go back to parent folder of d21 that is d1 and from d1 to main directory.