Add files via upload
We were given a valid parenthesis string and we had to remove the outermost parenthesis of each primitive substring. For example : s = (()())(()) in this string s, the primitive substrings would be (()()) and (()) and after removing their outermost parentheses the resultant string would be concatenated back as ()()(). A valid parenthesis string is a string which follows the following properties :
1. Every open bracket must be closed and at appropriate positions.
According to the constraints of the question: 1<= s.length <= 105 and s is a valid parenthesis string ????? If s.length == 1 how can s be a valid parenthesis string. Well anyways this didnt cause any problems regarding the submission of the solution so i guess its fine .