Add files via upload
So there are 2 approaches to this question, both of them are mentioned in the file; although they arent that different from one another its just a matter of which data structure you use and how you use it . Now regarding the question, we were to check the maximum nesting depth of the given valid parenthesis string. For example, Input: s = "(1+(2*3)+((8)/4))+1" in this example the maximum nesting depth is 3 of 8 ( 8 is inside of 3 brackets ).