Add files via upload
The problem was to remove duplicate adjacent characters in a string of lowercase english characters. For example s = "abbaca"; here it would go through the following iterations :
abbaca -> removing b: aaca -> removing a: ca -> no more duplicate adjacent characters .
The solution uses a stack; for each character in string s we check whether the last character in the stack that is stack[-1] is equal to the current character in string or not. For better time complexity use .join method as result = ''.join(stack) where '' is the separator.