Create FaultyKeyboard
You have a faulty keyboard in which upon typing character "i" the string existing before "i" is reversed. You are to return the final string after the whole string is typed . For example :
s = "string"
Output: "rtsng"
Explanation: 
After typing first character, the text on the screen is "s".
After the second character, the text is "st". 
After the third character, the text is "str".
Since the fourth character is an 'i', the text gets reversed and becomes "rts".
After the fifth character, the text is "rtsn". 
After the sixth character, the text is "rtsng". 
Therefore, we return "rtsng".